+++ app/cmd/recruit/other_game_recruit.js
@@ -1,5 +1,5 @@
-const { MessageEmbed, MessageActionRow, MessageButton } = require('discord.js');
-const { URLSearchParams } = require('url');
+const { MessageEmbed } = require('discord.js');
+const { recruitDeleteButton, recruitActionRow, disableButtons } = require('./button_components.js');
 
 module.exports = {
     otherGameRecruit: otherGameRecruit,
@@ -144,38 +144,4 @@ async function sendOtherGames(interaction, title, txt, color, image, logo) {
     } catch (error) {
         console.log(error);
     }
-}
-
-function recruitActionRow(msg, host_user) {
-    const joinParams = new URLSearchParams();
-    joinParams.append('d', 'jr');
-    joinParams.append('mid', msg.id);
-    joinParams.append('cid', msg.channel.id);
-    joinParams.append('hid', host_user.id);
-
-    const cancelParams = new URLSearchParams();
-    cancelParams.append('d', 'cr');
-    cancelParams.append('mid', msg.id);
-    cancelParams.append('cid', msg.channel.id);
-    cancelParams.append('hid', host_user.id);
-
-    const closeParams = new URLSearchParams();
-    closeParams.append('d', 'close');
-    closeParams.append('mid', msg.id);
-    closeParams.append('cid', msg.channel.id);
-    closeParams.append('hid', host_user.id);
-
-    return new MessageActionRow().addComponents([
-        new MessageButton().setCustomId(joinParams.toString()).setLabel('参加').setStyle('PRIMARY'),
-        new MessageButton().setCustomId(cancelParams.toString()).setLabel('キャンセル').setStyle('DANGER'),
-        new MessageButton().setCustomId(closeParams.toString()).setLabel('〆').setStyle('SECONDARY'),
-    ]);
-}
-function disableButtons() {
-    let buttons = new MessageActionRow().addComponents([
-        new MessageButton().setCustomId('join').setLabel('参加').setStyle('PRIMARY').setDisabled(),
-        new MessageButton().setCustomId('cancel').setLabel('キャンセル').setStyle('DANGER').setDisabled(),
-        new MessageButton().setCustomId('close').setLabel('〆').setStyle('SECONDARY').setDisabled(),
-    ]);
-    return buttons;
 }
